+++ e2e/tests/visual-output/rendering/index.spec.ts-snapshots/Rect-DropShadow-with-nonScaling-true.png
