+++ e2e/tests/visual-output/rendering/index.spec.ts-snapshots/Rect-with-strokeUniform-true-and-negative-scaling.png
