+++ e2e/tests/visual-output/rendering/index.spec.ts-snapshots/polygon-boundingbox-with-skew.png
